preparing for shortener auto deletion
Added logging for the time of creation of when the shortener was created. Will add logging for when the last time the actual shortener was used. 

How this works:
After one month, the shortener becomes available. 
*Being available doesnt make it so it can't be used. It still contains the link and can be used as normal. If used, it will just make it unavailable. for another month*
If it is in available, You can register a new link with that shortener. This will pretty much act the same as if you where creating a new shortener, however you are using a shortener that has been used before.